+++ app/src/main/java/com/example/stretchy/Screen.kt
@@ -5,4 +5,5 @@ sealed class Screen(val route: String) {
     object ExerciseCreatorScreen : Screen("exerciseCreatorScreen?id={id}&trainingType={trainingType}")
     object ExecuteTrainingScreen : Screen("executeTraining?id={id}")
     object TrainingListScreen : Screen("trainingListScreen")
+    object MetaTrainingScreen : Screen("metaTrainingScreen")
 }
