+++ app/src/main/java/com/example/stretchy/navigation/BottomNavScreen.kt
@@ -1,4 +1,4 @@
-package com.example.stretchy
+package com.example.stretchy.navigation
 
 import androidx.compose.ui.graphics.painter.Painter
